commit f6dea603a339e9ce32cc78b775543b0cccbefac2
    9955 loader: export boot pool and vdev GUID
commit 5b8f338ecee33a4471924e65d685fde98134bc31
    9787 snoop: add tftp OACK support
commit 7ec6bfcf4280022b95218f03e28fb50217e4b6ff
    9902 mac: mac_soft_ring_poll should use size_t
commit 124619e5421ff751660c822c15c320ddf7800a5a
    9954 loader: always set media size from partition.

+++ usr/src/uts/common/io/mac/mac_soft_ring.c
@@ -152,8 +152,8 @@ mac_soft_ring_create(int id, clock_t wait, uint16_t type,
     processorid_t cpuid, mac_direct_rx_t rx_func, void *x_arg1,
     mac_resource_handle_t x_arg2)
 {
-	mac_soft_ring_t 	*ringp;
-	char 			name[S_RING_NAMELEN];
+	mac_soft_ring_t		*ringp;
+	char			name[S_RING_NAMELEN];
 
 	bzero(name, 64);
 	ringp = kmem_cache_alloc(mac_soft_ring_cache, KM_SLEEP);
@@ -353,7 +353,7 @@ mac_rx_soft_ring_drain(mac_soft_ring_t *ringp)
 	mblk_t		*mp;
 	void		*arg1;
 	mac_resource_handle_t arg2;
-	timeout_id_t 	tid;
+	timeout_id_t	tid;
 	mac_direct_rx_t	proc;
 	size_t		sz;
 	int		cnt;
@@ -532,7 +532,7 @@ mac_soft_ring_intr_disable(void *arg)
  * setup.
  */
 mblk_t *
-mac_soft_ring_poll(mac_soft_ring_t *ringp, int bytes_to_pickup)
+mac_soft_ring_poll(mac_soft_ring_t *ringp, size_t bytes_to_pickup)
 {
 	mblk_t	*head, *tail;
 	mblk_t	*mp;
@@ -644,10 +644,10 @@ mac_soft_ring_signal(mac_soft_ring_t *softring, uint_t sr_flag)
 static void
 mac_tx_soft_ring_drain(mac_soft_ring_t *ringp)
 {
-	mblk_t 			*mp;
-	void 			*arg1;
-	void 			*arg2;
-	mblk_t 			*tail;
+	mblk_t			*mp;
+	void			*arg1;
+	void			*arg2;
+	mblk_t			*tail;
 	uint_t			saved_pkt_count, saved_size;
 	mac_tx_stats_t		stats;
 	mac_soft_ring_set_t	*mac_srs = ringp->s_ring_set;
